Rename FlowState to GenerationState
